Specifically call out --help for ssh command
It took me a long time to discover the new ways to specify gateway host and user, etc.

If you are going to start supporting `--help` for most commands, trash this PR. If you're not going to get there soon, this little addition could help save folks some time. I discovered the `--help` option by reading the source code for the ssh command.