Update manifest to v2
We're just following some onboarding docs and there was a discrepancy between the v2 example and these docs which seem to be out of date. I'm not sure if this change should be made or if we should be linking to a separate v2 docs page in our onboarding.

Also unsure if all references to "job(s)" should now be references to "instance group(s)" or not.