Add steps for expired blobstore CA replacement
Essentially the same steps for NATS CA. It's likely that all of them
will have expired, so operators can reference both this page and the
NATS CA page to combine the steps. Some operators adjust the duration of
some CAs and not others, so NATS CA rotation may not be applicable
during blobstore CA rotation.