Update cli-v2-install.md
The brew command line seems to be wrong. The correct folder is "homebrew-tap" instead of only "tap".
When using the original command you will be ask for the GitHub credentials, but an error will occur even after you typed in a valid user and password.
I tried it on my iMac and the corrected command line worked.