Remove mention of permit nats access
we pulled the nats FW implemention from the stemcell builder since it didn't work properly across IaaSs. 
When reverting the changes in the stemcell builder, we forgot to remove this section from the docs.  
These additional debugging steps relate to a script that was reverted from the stemcell builder. Currently
there is work around re-adding the FW functionality via the bosh-agent. This section will be re-updated
with the proper steps once the stemcell containing the new agent versions is released.