Remove unneccessary vars generation
In the second ops file, the /variables do not need to be touched. The only 2 cases where they could be not there would be:
- You run Step3 before step 1
- You forget to remove the ops after step 5

This way in both cases the bosh deploy errors and will not cause all VMs to lose connection to BOSH. We internally use it like this for quite a while now and it prevented some incidents already so I think everyone should have that.