Add practical examples of bosh handling AZ changes
[ #183715145 ]

The behavior of bosh in regards to adding / removing / replacing AZs to
a deployment and the subsequent rebalancing of VMs (especially VMs with
persistent disks) is somewhat hard to understand. This adds examples for
the expected behavior of bosh when attempting to change the AZs for
instance_groups in several scenarios.